requirements.txt
# Retrieval
arxiv==2.1.0
semanticscholar==0.8.4
rank_bm25==0.2.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1

# LLM
cerebras-cloud-sdk>=1.0.0

# NLP / Embeddings / Transformers
transformers==4.40.0
torch==2.3.0
sentence-transformers==2.7.0
spacy==3.8.4
datasets==2.19.0
accelerate==0.30.0

# Clustering & Math
umap-learn==0.5.6
hdbscan==0.8.33
scikit-learn==1.4.2
scipy==1.13.0
numpy==1.26.4

# Graph
networkx==3.3
pyvis==0.3.2

# Data / Viz
pandas==2.2.2
matplotlib==3.9.0
plotly==5.22.0
kaleido==0.2.1

# API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic==2.7.1
httpx==0.27.0
python-dotenv==1.0.1
aiofiles==23.2.1
